BREAKING CHANGE: Follow background option as it's set
Before this change, the option was followed once
on initial load due to `require` behaviour.

Additionally, theme option is now respected
on repeated loads. E.g.:

```
:lua require('onenord').setup({theme = 'dark'})
:lua require('onenord').setup({theme = 'light'})
```

Yields a light variant, as you'd expect.

In order to unset forced theme, set it to `false`.
This is due to the fact that `{a = nil}` and `{}`
are equivalent. As such, providing `theme = nil`
is equivalent to not providing it at all.